# contracts/openapi.yaml
# OpenAPI 3.1 document (JSON Schema 2020-12 dialect). Single source of truth for the contract.
# Objects use additionalProperties: false. Refer to a schema with "#/components/schemas/<Name>".
openapi: 3.1.0
info:
  title: Anioł Stróż contract
  version: 0.0.1
  description: Data contract shared by backend (Java), frontend (Angular) and Android. Paths cover /api/demo/* so far; other endpoints are added task by task (CON-02). StageHits is snake_case on purpose (Claude response); everything else is camelCase.
paths:
  /api/demo/scenarios:
    get:
      operationId: listScenarios
      tags:
        - demo
      summary: List the demo scenarios available for replay.
      responses:
        "200":
          description: Scenarios in file-name order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScenarioSummary"
  /api/demo/replay:
    post:
      operationId: startReplay
      tags:
        - demo
      summary: Start replaying a scenario as a call. Only SCRIPTED is supported for now.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplayRequest"
      responses:
        "202":
          description: The call has started and segments are being played.
        "400":
          description: Unsupported mode or invalid body.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Unknown scenarioId.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Another call is already active (BE-04).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/demo/stop:
    post:
      operationId: stopReplay
      tags:
        - demo
      summary: Stop the running replay and end its call. Does nothing if none is running.
      responses:
        "204":
          description: No active replay remains.
components:
  schemas:
    Alert:
      title: Alert
      description: Alert created by deterministic code. Texts come from a pre-written template, never from the model.
      type: object
      additionalProperties: false
      required:
        - alertId
        - callId
        - level
        - stages
        - templateId
        - shortText
        - advice
        - triggeredBy
        - createdAt
        - mode
      properties:
        alertId:
          type: string
          minLength: 1
        callId:
          type: string
          minLength: 1
        level:
          $ref: "#/components/schemas/RiskLevel"
        stages:
          type: array
          items:
            $ref: "#/components/schemas/StageHit"
          description: Validated hits only.
        templateId:
          type: string
          minLength: 1
        shortText:
          type: string
          minLength: 1
          description: One-sentence reason from the template, shown and spoken to the senior.
        advice:
          type: string
          minLength: 1
          description: Advice text from the template.
        triggeredBy:
          type: string
          enum:
            - llm
            - keywords
            - both
        createdAt:
          type: string
          format: date-time
        mode:
          $ref: "#/components/schemas/Mode"
    Decision:
      title: Decision
      description: A decision made by a person about an alert.
      type: object
      additionalProperties: false
      required:
        - alertId
        - actor
        - decision
        - at
      properties:
        alertId:
          type: string
          minLength: 1
        actor:
          type: string
          enum:
            - senior
            - family
        decision:
          type: string
          enum:
            - hung_up
            - called_trusted
            - false_alarm
            - confirmed_scam
        at:
          type: string
          format: date-time
    EventEnvelope:
      title: EventEnvelope
      description: "Every message on /ws/events: the type selects the payload, mode is always present."
      oneOf:
        - title: TranscriptSegmentEvent
          type: object
          additionalProperties: false
          required:
            - type
            - mode
            - at
            - payload
          properties:
            type:
              const: transcript.segment
            mode:
              $ref: "#/components/schemas/Mode"
            at:
              type: string
              format: date-time
            payload:
              $ref: "#/components/schemas/TranscriptSegment"
        - title: RiskUpdateEvent
          type: object
          additionalProperties: false
          required:
            - type
            - mode
            - at
            - payload
          properties:
            type:
              const: risk.update
            mode:
              $ref: "#/components/schemas/Mode"
            at:
              type: string
              format: date-time
            payload:
              $ref: "#/components/schemas/RiskUpdate"
        - title: AlertCreatedEvent
          type: object
          additionalProperties: false
          required:
            - type
            - mode
            - at
            - payload
          properties:
            type:
              const: alert.created
            mode:
              $ref: "#/components/schemas/Mode"
            at:
              type: string
              format: date-time
            payload:
              $ref: "#/components/schemas/Alert"
        - title: AlertDecisionEvent
          type: object
          additionalProperties: false
          required:
            - type
            - mode
            - at
            - payload
          properties:
            type:
              const: alert.decision
            mode:
              $ref: "#/components/schemas/Mode"
            at:
              type: string
              format: date-time
            payload:
              $ref: "#/components/schemas/Decision"
        - title: SystemStatusEvent
          type: object
          additionalProperties: false
          required:
            - type
            - mode
            - at
            - payload
          properties:
            type:
              const: system.status
            mode:
              $ref: "#/components/schemas/Mode"
            at:
              type: string
              format: date-time
            payload:
              $ref: "#/components/schemas/SystemStatus"
        - title: CallStartedEvent
          type: object
          additionalProperties: false
          required:
            - type
            - mode
            - at
            - payload
          properties:
            type:
              const: call.started
            mode:
              $ref: "#/components/schemas/Mode"
            at:
              type: string
              format: date-time
            payload:
              type: object
              additionalProperties: false
              required:
                - callId
              properties:
                callId:
                  type: string
                  minLength: 1
        - title: CallEndedEvent
          type: object
          additionalProperties: false
          required:
            - type
            - mode
            - at
            - payload
          properties:
            type:
              const: call.ended
            mode:
              $ref: "#/components/schemas/Mode"
            at:
              type: string
              format: date-time
            payload:
              type: object
              additionalProperties: false
              required:
                - callId
                - hadAlert
              properties:
                callId:
                  type: string
                  minLength: 1
                hadAlert:
                  type: boolean
    Mode:
      title: Mode
      description: Honest operating mode carried by every event, audit record and screen.
      type: string
      enum:
        - LIVE
        - REPLAY
        - SCRIPTED
        - MOCK
    RiskLevel:
      title: RiskLevel
      description: Discrete risk level computed deterministically by the backend.
      type: string
      enum:
        - none
        - low
        - medium
        - high
    RiskUpdate:
      title: RiskUpdate
      description: Current deterministic risk level of a call. Shown in words, never as a percentage.
      type: object
      additionalProperties: false
      required:
        - callId
        - level
        - previousLevel
        - stages
        - warningSigns
      properties:
        callId:
          type: string
          minLength: 1
        level:
          $ref: "#/components/schemas/RiskLevel"
        previousLevel:
          $ref: "#/components/schemas/RiskLevel"
        stages:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/StageId"
          description: Distinct stages counted so far (validated hits only).
        warningSigns:
          type: integer
          minimum: 0
          description: Number of distinct validated stages.
    ProblemDetail:
      title: ProblemDetail
      description: RFC 9457 error body. detail is Polish text for the UI (API-06).
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
    ReplayRequest:
      title: ReplayRequest
      description: Body of POST /api/demo/replay.
      type: object
      additionalProperties: false
      required:
        - scenarioId
        - mode
      properties:
        scenarioId:
          type: string
          pattern: ^[a-z0-9-]+$
        mode:
          $ref: "#/components/schemas/Mode"
        speed:
          type: number
          format: double
          minimum: 0.1
          maximum: 20
          default: 1
          description: Playback speed multiplier; each segment waits delayMs / speed.
    Scenario:
      title: Scenario
      description: A named demo scenario fed to the pipeline in REPLAY, SCRIPTED or MOCK mode.
      type: object
      additionalProperties: false
      required:
        - scenarioId
        - title
        - description
        - mode
        - segments
      properties:
        scenarioId:
          type: string
          pattern: ^[a-z0-9-]+$
        title:
          type: string
          minLength: 1
        description:
          type: string
        mode:
          $ref: "#/components/schemas/Mode"
        audioFile:
          type: string
          minLength: 1
          description: Team recording under recordings/, REPLAY only.
        source:
          type: string
          minLength: 1
          description: Provenance of the synthetic script.
        sourceUrl:
          type: string
          minLength: 1
          description: Public description of the scam the script is based on, if known.
        expected:
          type: object
          description: Expected properties for the evaluation run, not scores.
          additionalProperties: false
          required:
            - maxLevel
            - mustHitStages
            - mustNotHitStages
          properties:
            maxLevel:
              $ref: "#/components/schemas/RiskLevel"
            mustHitStages:
              type: array
              uniqueItems: true
              items:
                $ref: "#/components/schemas/StageId"
            mustNotHitStages:
              type: array
              uniqueItems: true
              items:
                $ref: "#/components/schemas/StageId"
        segments:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required:
              - speaker
              - text
              - delayMs
            properties:
              speaker:
                type: string
                enum:
                  - A
                  - B
                  - unknown
              text:
                type: string
                minLength: 1
              delayMs:
                type: integer
                minimum: 0
    ScenarioSummary:
      title: ScenarioSummary
      description: Listing entry for GET /api/demo/scenarios.
      type: object
      additionalProperties: false
      required:
        - scenarioId
        - title
        - description
      properties:
        scenarioId:
          type: string
          pattern: ^[a-z0-9-]+$
        title:
          type: string
        description:
          type: string
    Sensitivity:
      title: Sensitivity
      description: Household sensitivity setting.
      type: string
      enum:
        - calm
        - standard
        - sensitive
    Settings:
      title: Settings
      description: Household settings. Never sent to Claude or STT.
      type: object
      additionalProperties: false
      required:
        - seniorConsent
        - familyConsent
        - contacts
        - sensitivity
        - retentionDays
      properties:
        seniorConsent:
          type: boolean
        familyConsent:
          type: boolean
        contacts:
          type: array
          maxItems: 5
          items:
            type: object
            additionalProperties: false
            required:
              - name
              - phone
            properties:
              name:
                type: string
                minLength: 1
                maxLength: 100
              phone:
                type: string
                pattern: ^\+?[0-9 ]{9,15}$
        sensitivity:
          $ref: "#/components/schemas/Sensitivity"
        retentionDays:
          type: integer
          minimum: 1
          maximum: 90
    StageHit:
      title: StageHit
      description: "Evidence of one manipulation stage: a verbatim quote from one transcript segment."
      type: object
      additionalProperties: false
      required:
        - stage
        - segId
        - quote
        - speakerRole
        - source
        - validated
      properties:
        stage:
          $ref: "#/components/schemas/StageId"
        segId:
          type: string
          pattern: ^s[0-9]+$
        quote:
          type: string
          minLength: 1
        speakerRole:
          type: string
          enum:
            - caller
            - senior
            - background
            - unclear
        source:
          type: string
          enum:
            - llm
            - keywords
        validated:
          type: boolean
          description: Result of QuoteValidator.
    StageHits:
      title: StageHits
      description: "Claude structured output: evidence only. Snake_case on purpose, this is the model response schema."
      type: object
      additionalProperties: false
      required:
        - stage_hits
      properties:
        stage_hits:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - stage
              - segment_id
              - quote
              - speaker_role
            properties:
              stage:
                $ref: "#/components/schemas/StageId"
              segment_id:
                type: string
              quote:
                type: string
              speaker_role:
                type: string
                enum:
                  - caller
                  - senior
                  - background
                  - unclear
    StageId:
      title: StageId
      description: Manipulation stage recognised in a call.
      type: string
      enum:
        - AUTHORITY_CLAIM
        - URGENT_THREAT
        - SECRECY_DEMAND
        - ISOLATION
        - MONEY_REQUEST
        - PAYMENT_CHANNEL
        - REMOTE_ACCESS
        - PERSONAL_DATA_REQUEST
    SystemStatus:
      title: SystemStatus
      description: State of one component. Every failure is published, never silent.
      type: object
      additionalProperties: false
      required:
        - component
        - state
        - message
        - at
      properties:
        component:
          type: string
          enum:
            - audio
            - stt
            - ai
            - backend
        state:
          type: string
          enum:
            - ok
            - degraded
            - down
        message:
          type: string
          description: Polish text for the UI.
        at:
          type: string
          format: date-time
    TranscriptSegment:
      title: TranscriptSegment
      description: One interim or final segment of the call transcript.
      type: object
      additionalProperties: false
      required:
        - callId
        - segId
        - tStartMs
        - tEndMs
        - text
        - isFinal
        - speaker
      properties:
        callId:
          type: string
          minLength: 1
        segId:
          type: string
          pattern: ^s[0-9]+$
          description: Segment id such as s12.
        tStartMs:
          type: integer
          minimum: 0
        tEndMs:
          type: integer
          minimum: 0
        text:
          type: string
        isFinal:
          type: boolean
        speaker:
          type: string
          enum:
            - A
            - B
            - unknown
          description: Only meaningful if the STT provider diarizes.
        sttConfidence:
          type: number
          minimum: 0
          maximum: 1
          description: Only present if the STT provider returns it.